Refactor and centralize texture loading methods for PNG files and data, along with metadata extraction, into `MeshUtils`.
NP-1234
