+++ src/communication/ArcusCommunicationPrivate.cpp
@@ -18,6 +18,7 @@
 #include "Slice.h"
 #include "settings/types/LayerIndex.h"
 #include "utils/Matrix4x3D.h" //To convert vertices to integer-points.
+#include "utils/MeshUtils.h"
 #include "utils/Point3F.h" //To accept vertices (which are provided in floating point).
 
 namespace cura
@@ -163,125 +164,7 @@ void ArcusCommunication::Private::readMeshGroupMessage(const proto::ObjectList& 
 
 void ArcusCommunication::Private::loadTextureData(const std::string& texture_str, Mesh& mesh)
 {
-    if (texture_str.empty())
-    {
-        return;
-    }
-
-    auto texture_data = reinterpret_cast<const unsigned char*>(texture_str.data());
-    const size_t texture_size = texture_str.size();
-    png_image raw_texture = {};
-    raw_texture.version = PNG_IMAGE_VERSION;
-    if (! png_image_begin_read_from_memory(&raw_texture, texture_data, texture_size))
-    {
-        spdlog::error("Error when beginning reading mesh texture: {}", raw_texture.message);
-        return;
-    }
-
-    std::vector<uint8_t> buffer(PNG_IMAGE_SIZE(raw_texture));
-    if (! png_image_finish_read(&raw_texture, nullptr, buffer.data(), 0, nullptr) || buffer.empty())
-    {
-        spdlog::error("Error when finishing reading mesh texture: {}", raw_texture.message);
-        return;
-    }
-
-    // Make sure pointer will be destroyed when leaving
-    std::unique_ptr<png_struct, void (*)(png_structp)> png_ptr(
-        png_create_read_struct(PNG_LIBPNG_VER_STRING, nullptr, nullptr, nullptr),
-        [](png_structp png_ptr_destroy)
-        {
-            png_destroy_read_struct(&png_ptr_destroy, nullptr, nullptr);
-        });
-    if (! png_ptr)
-    {
-        return;
-    }
-
-    // Make sure pointer will be destroyed when leaving
-    std::unique_ptr<png_info, void (*)(png_infop)> info_ptr(
-        png_create_info_struct(png_ptr.get()),
-        [](png_infop info_ptr_destroy)
-        {
-            png_destroy_read_struct(nullptr, &info_ptr_destroy, nullptr);
-        });
-    if (! info_ptr)
-    {
-        return;
-    }
-
-    if (setjmp(png_jmpbuf(png_ptr.get())) != 0)
-    {
-        return;
-    }
-
-    struct PngReadContext
-    {
-        const unsigned char* data;
-        size_t size;
-        size_t offset;
-    } read_context{ texture_data, texture_size, 0 };
-
-    png_set_read_fn(
-        png_ptr.get(),
-        &read_context,
-        [](const png_structp read_png_ptr, const png_bytep out_bytes, const png_size_t byte_count_to_read)
-        {
-            auto* context = static_cast<PngReadContext*>(png_get_io_ptr(read_png_ptr));
-            if (context->offset + byte_count_to_read > context->size)
-            {
-                png_error(read_png_ptr, "Read beyond end of buffer");
-            }
-            memcpy(out_bytes, context->data + context->offset, byte_count_to_read);
-            context->offset += byte_count_to_read;
-        });
-    png_read_info(png_ptr.get(), info_ptr.get());
-
-    png_textp text_ptr;
-    int num_text;
-    if (png_get_text(png_ptr.get(), info_ptr.get(), &text_ptr, &num_text) <= 0)
-    {
-        return;
-    }
-
-    auto texture_data_mapping = std::make_shared<TextureDataMapping>();
-    for (int i = 0; i < num_text; ++i)
-    {
-        if (std::string(text_ptr[i].key) == "Description")
-        {
-            rapidjson::MemoryStream json_memory_stream(text_ptr[i].text, text_ptr[i].text_length);
-
-            rapidjson::Document json_document;
-            json_document.ParseStream(json_memory_stream);
-            if (json_document.HasParseError())
-            {
-                spdlog::error("Error parsing texture data mapping (offset {}): {}", json_document.GetErrorOffset(), GetParseError_En(json_document.GetParseError()));
-                return;
-            }
-
-            for (auto it = json_document.MemberBegin(); it != json_document.MemberEnd(); ++it)
-            {
-                std::string feature_name = it->name.GetString();
-
-                const rapidjson::Value& array = it->value;
-                if (array.IsArray() && array.Size() == 2)
-                {
-                    (*texture_data_mapping)[feature_name] = TextureBitField{ array[0].GetUint(), array[1].GetUint() };
-                }
-            }
-
-            break;
-        }
-    }
-
-    if (! texture_data_mapping->empty())
-    {
-        mesh.texture_ = std::make_shared<Image>(
-            raw_texture.width,
-            raw_texture.height,
-            PNG_IMAGE_SAMPLE_COMPONENT_SIZE(raw_texture.format) * PNG_IMAGE_SAMPLE_CHANNELS(raw_texture.format),
-            std::move(buffer));
-        mesh.texture_data_mapping_ = texture_data_mapping;
-    }
+    MeshUtils::loadTextureFromString(texture_str, mesh);
 }
 
 } // namespace cura
